Multi-modifier protection (MMP)
Add a new thread, mmp_thread, started at import/open (like sync_thread).
The new thread writes uberblocks to the last slot in each label, at a
set frequency, to indicate to other hosts the pool is actively imported.
These uberblocks are the last synced uberblock with an updated
timestamp.

Add a new optional uberblock field, mmp_delay, which reflects the time
between mmp writes.

During import, read uberblocks repeatedly, checking for change in the
selected uberblock.  Fail with EBUSY if such change is found.  Fail with
EINTR if the user presses Ctrl-C.

Allow the user to control the period between MMP writes, and the
duration of the activity test on import, via a new module parameter
zfs_mmp_interval.  The period is specified in milliseconds.  The
activity test duration is calculated from this value, and from the
mmp_delay in the "best" uberblock found initially.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>