MSEC2NSEC and ZPOOL_CONFIG_IMPORT_TXG fixes
** mmp_interval is initialized as MSEC2NSEC(zfs_mmp_interval) and is
updated using the same conversion.

In two places, the conversion was performed a second time accidentally,
like
	next_time = start + MSEC2NSEC(mmp_interval)

with the result that zfs_mmp_interval was effectively multiplied by an
additional factor of 1000 unintentionally.

** If the tryimport added key ZPOOL_CONFIG_IMPORT_TXG to the config, and
its value is 0, activity was detected.

No need to execute the activity check again, just fail the import
immediately.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>